requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
autologin==0.1.4
Automat==0.7.0
Brotli==1.0.7
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
crochet==1.9.0
cryptography==2.3.1
cssselect==1.0.3
device-detector==0.3
dj-database-url==0.5.0
Django==2.1.3
django-widget-tweaks==1.4.3
django-zxcvbn-password==2.0.2
django-zxcvbn-password-validator==1.2.5
docopt==0.6.2
Flask==0.10.1
Flask-Admin==1.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
formasaurus==0.8.1
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
numpy==1.15.4
parsel==1.5.1
pinax-points==1.0
py-cpuinfo==4.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-crfsuite==0.9.6
pytz==2018.7
PyYAML==3.13
queuelib==1.5.0
regex==2018.11.7
requests==2.20.0
requests-file==1.4.3
scikit-learn==0.20.0
scipy==1.1.0
Scrapy==1.5.1
scrapy-splash==0.7.2
selenium==3.141.0
service-identity==17.0.0
six==1.11.0
sklearn-crfsuite==0.3.6
SQLAlchemy==1.2.13
tabulate==0.8.2
tldextract==2.2.0
tqdm==4.28.1
Twisted==18.9.0
urllib3==1.24.1
w3lib==1.19.0
Werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.10.11
WTForms==2.1
zope.interface==4.6.0
zxcvbn==4.4.27
zxcvbn-python==4.4.24
